backend/requirements.txt
alpaca-py
lumibot
pandas
pandas-ta
python-dotenv
fastapi
uvicorn
google-generativeai
google-genai
